sharing logic between different download spots
